Remove a bunch of the symmetries by not rotating `dingusS`
Solutions that differ from each other only by a rotation or a
reflection are not really distinct. (There are 48
rotations/reflections of a cube.)

It's hard to eliminate all of the degeneracy, but we can reduce it by
a big factor by only trying to place the `dingusS` piece in a single
orientation. This doesn't eliminate all of the degeneracy because the
`dingusS` piece itself has fourfold symmetry (reflection through its
plane and 180° within its plane.)